requirements.txt
scapy
flask
flask-socketio
eventlet
python-dotenv
pyyaml
requests
netifaces